Set up two new top-level tabs with Details and Contributors
