+++ client/containers/Pub/SpubHeader/DetailsTab.tsx
@@ -3,7 +3,7 @@ import { Label, InputGroup } from '@blueprintjs/core';
 import { useDebouncedCallback } from 'use-debounce';
 
 import { PubPageData, Pub, DocJson } from 'types';
-import { MinimalEditor } from 'components';
+import { MinimalEditor, DownloadChooser } from 'components';
 
 type Props = {
 	pub: Pub;
@@ -12,14 +12,39 @@ type Props = {
 	onUpdateAbstract: (abstract: DocJson) => Promise<unknown>;
 };
 
-const TitleDescriptionAbstract = (props: Props) => {
+const DetailsTab = (props: Props) => {
 	const [onUpdatePubDebounced] = useDebouncedCallback(props.onUpdatePub, 250);
+	const renderDefaultDownloadFile = () => {
+		return (
+			<div className="submission-tab-prompt">
+				<div className="submission-tab-prompt-text">
+					<h2>Default Download File</h2>
+					<p className="submission-tab-prompt-text">
+						You can upload a file, like a PDF with custom styling, to associate with
+						this submission
+					</p>
+				</div>
+
+				<div>
+					<DownloadChooser
+						pubData={props.pub}
+						communityId={props.pub.communityId}
+						onSetDownloads={props.onUpdatePub}
+						text="UPLOAD YOUR FILE"
+						isMinimal
+						isSmall
+					/>
+				</div>
+			</div>
+		);
+	};
+
 	return (
 		<>
 			<span className="info-text">
-				The information you enter in this form and pub body below will be used to create a
-				submission pub, which you can preview at any time before making your final
-				submission.
+				The information you enter in this form and the content section below will be used to
+				create your submission. You can use the Preview tab to see how your submission will
+				appear.
 			</span>
 			<Label htmlFor="pub-title-input">
 				<h2>Title of your submission pub</h2>
@@ -30,13 +55,13 @@ const TitleDescriptionAbstract = (props: Props) => {
 						onUpdatePubDebounced({ title: evt.target.value })
 					}
 					defaultValue={props.pub.title}
-					placeholder="Enter pub title here..."
+					placeholder="Type your submission's title here..."
 				/>
 			</Label>
 			<Label htmlFor="abstract-input">
 				<h2>Abstract</h2>
 				<MinimalEditor
-					placeholder="Enter abstract here..."
+					placeholder="Type your submission's abstract here..."
 					initialContent={props.abstract}
 					onEdit={(doc) => props.onUpdateAbstract(doc.toJSON() as DocJson)}
 					debounceEditsMs={300}
@@ -46,25 +71,11 @@ const TitleDescriptionAbstract = (props: Props) => {
 				/>
 			</Label>
 			<Label>
-				<h2>Description</h2>
-				<InputGroup
-					className="submission-input"
-					placeholder="Enter description text here..."
-					defaultValue={props.pub.description}
-					onChange={(evt: React.ChangeEvent<HTMLInputElement>) =>
-						onUpdatePubDebounced({ description: evt.target.value })
-					}
-				/>
-			</Label>
-			<Label>
 				<h2>Pub Content</h2>
-				<span className="input-prompt">
-					Enter your primary submission content in the pub body below by typing or by
-					importing files in which you already have content prepared.
-				</span>
 			</Label>
+			{renderDefaultDownloadFile()}
 		</>
 	);
 };
 
-export default TitleDescriptionAbstract;
+export default DetailsTab;
