+++ client/containers/Pub/SpubHeader/SubmissionTab/SpubSettings.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { PubPageData, DefinitelyHas } from 'types';
 
 import { usePageContext } from 'utils/hooks';
-import { PubThemePicker, PopoverButton, DownloadChooser } from 'components';
+import { PubThemePicker, PopoverButton } from 'components';
 import SmallHeaderButton from '../../PubHeader/SmallHeaderButton';
 
 require('./spubSettings.scss');
@@ -51,42 +51,7 @@ const SpubSettings = (props: Props) => {
 		);
 	};
 
-	const renderDefaultDownloadFile = () => {
-		return (
-			<div className="submission-tab-prompt">
-				<div className="submission-tab-prompt-text">
-					<h2>Default Download File</h2>
-					<p className="submission-tab-prompt-text">
-						Readers of your submission pub can choose to download it as a file,
-						available in different formats automatically generated by PubPub.
-						<br />
-						<br />
-						You can upload a file, like a PDF with custom styling, to associate with
-						your pub. It will be provided to readers as the pub's default download, but
-						they'll still be able to use the automatic export tools.
-					</p>
-				</div>
-
-				<div>
-					<DownloadChooser
-						pubData={props.pubData}
-						communityId={communityData.id}
-						onSetDownloads={props.onUpdatePub}
-						text="UPLOAD YOUR FILE"
-						isMinimal
-						isSmall
-					/>
-				</div>
-			</div>
-		);
-	};
-
-	return (
-		<div className="submission-tab-tabs">
-			{renderHeaderAndBackgroundSetting()}
-			{renderDefaultDownloadFile()}
-		</div>
-	);
+	return <div className="submission-tab-tabs">{renderHeaderAndBackgroundSetting()}</div>;
 };
 
 export default SpubSettings;
